Adding some updates in order to improve the API

+++ src/main/java/com/matancita/yngreni/web/PagareController.java
@@ -1,11 +1,14 @@
 package com.matancita.yngreni.web;
 
+import com.matancita.yngreni.DTO.Messages;
 import com.matancita.yngreni.domain.Pagare;
+import com.matancita.yngreni.domain.RecibosGen;
 import com.matancita.yngreni.service.PagareService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.time.LocalDateTime;
 import java.util.List;
 
 @RestController
@@ -15,11 +18,13 @@ public class PagareController {
     @Autowired
     PagareService pagareService;
 
-    @PostMapping("/save")
-    public ResponseEntity<?> savePagare(@RequestBody List<Pagare> pagare){
-        pagareService.insertAll(pagare);
-        return ResponseEntity.ok(pagare);
-    }
+   // @PostMapping("/save")
+   // public ResponseEntity<?> savePagare(@RequestBody List<Pagare> pagare){
+    //    pagareService.insertAll(pagare);
+     //   return ResponseEntity.ok(pagare);
+    //}
+
+
 
     @GetMapping("/list")
     public ResponseEntity<?> listPagares() {
@@ -40,19 +45,42 @@ public class PagareController {
     @PutMapping("/update/{id}")
     public ResponseEntity<?> updatePagare(@PathVariable Long id, @RequestBody Pagare pagare){
         Pagare pagareExist = pagareService.getById(id);
-        pagareExist.setNoPagare(pagare.getNoPagare());
-        pagareExist.setCapital(pagare.getCapital());
-        pagareExist.setInteres(pagare.getInteres());
-        pagareExist.setTotal(pagare.getTotal());
-        pagareExist.setVencimiento(pagare.getVencimiento());
-        pagareService.update(pagareExist);
-        return  ResponseEntity.ok(pagareExist);
+        if(pagareExist!=null){
+            pagareExist.setNoPagare(pagare.getNoPagare());
+            pagareExist.setCapital(pagare.getCapital());
+            pagareExist.setInteres(pagare.getInteres());
+            pagareExist.setTotal(pagare.getTotal());
+            pagareExist.setVencimiento(pagare.getVencimiento());
+            pagareService.update(pagareExist);
+            return  ResponseEntity.ok(pagareExist);
+        }else{
+            return ResponseEntity.notFound().build();
+        }
     }
 
+    @PutMapping("/printpagare/{id}")
+    public ResponseEntity<?> printPagare(@PathVariable Long id){
+        Pagare pagareToPrint = pagareService.getById(id);
+        if(pagareToPrint!=null){
+            RecibosGen recibosGen = new RecibosGen();
+            recibosGen.setFecha(LocalDateTime.now());
+            recibosGen.setValor(pagareToPrint.getTotal());
+            pagareToPrint.setReciboGen(recibosGen);
+            pagareService.update(pagareToPrint);
+            return ResponseEntity.ok(pagareToPrint);
+        }else{
+            return ResponseEntity.notFound().build();
+        }
+    }
     @DeleteMapping("/delete/{id}")
     public ResponseEntity<?> deletepagare(@PathVariable Long id){
         Pagare pagare = pagareService.getById(id);
-        pagareService.delete(pagare);
-        return ResponseEntity.ok("Pagare Eliminado");
+        if(pagare!=null){
+            pagareService.delete(pagare);
+            return ResponseEntity.ok(new Messages("Pagare eliminado"));
+        }else{
+            return ResponseEntity.notFound().build();
+        }
+
     }
 }
